refactor: extract common event processing logic
- Move event processing logic to common package:
  • ProcessTransfers
  • ProcessCreatedOrders
  • ProcessSettledOrders
  • ProcessRefundedOrders
- Remove duplicate code from EVM and Tron indexers
- Add block time to test fixtures
- Update test assertions for Engine integration

This change:
- Reduces code duplication
- Improves maintainability
- Makes event processing consistent across networks
- Prepares test fixtures for Engine integration
