+++ services/indexer/evm.go
@@ -4,18 +4,13 @@ import (
 	"context"
 	"fmt"
 	"strings"
-	"sync"
 
 	"github.com/paycrest/aggregator/ent"
-	"github.com/paycrest/aggregator/ent/lockpaymentorder"
-	"github.com/paycrest/aggregator/ent/paymentorder"
 	"github.com/paycrest/aggregator/services"
 	"github.com/paycrest/aggregator/services/common"
 	"github.com/paycrest/aggregator/services/order"
-	"github.com/paycrest/aggregator/storage"
 	"github.com/paycrest/aggregator/types"
 	"github.com/paycrest/aggregator/utils"
-	"github.com/paycrest/aggregator/utils/logger"
 	"github.com/shopspring/decimal"
 )
 
@@ -86,14 +81,9 @@ func (s *IndexerEVM) IndexTransfer(ctx context.Context, rpcClient types.RPCClien
 		return nil
 	}
 
-	// Process receive addresses and update their status
-	if err := common.ProcessReceiveAddresses(ctx, s.order, s.priorityQueue, unknownAddresses, addressToEvent); err != nil {
-		return err
-	}
-
-	// Process linked addresses and create payment orders
-	if err := common.ProcessLinkedAddresses(ctx, s.order, unknownAddresses, addressToEvent, token); err != nil {
-		return err
+	err = common.ProcessTransfers(ctx, s.order, s.priorityQueue, unknownAddresses, addressToEvent, token)
+	if err != nil {
+		return fmt.Errorf("IndexTransfer.processTransfers: %w", err)
 	}
 
 	return nil
@@ -143,74 +133,10 @@ func (s *IndexerEVM) IndexOrderCreated(ctx context.Context, rpcClient types.RPCC
 		return nil
 	}
 
-	orders, err := storage.Client.PaymentOrder.
-		Query().
-		Where(paymentorder.TxHashIn(txHashes...)).
-		WithToken(func(tq *ent.TokenQuery) {
-			tq.WithNetwork()
-		}).
-		WithRecipient().
-		All(ctx)
+	err = common.ProcessCreatedOrders(ctx, network, txHashes, hashToEvent, s.order, s.priorityQueue)
 	if err != nil {
-		logger.Infof("IndexOrderCreated.fetchOrders: %v", err)
-		return fmt.Errorf("IndexOrderCreated.fetchOrders: %w", err)
+		return fmt.Errorf("IndexOrderCreated.processCreatedOrders: %w", err)
 	}
-
-	var wg sync.WaitGroup
-	for _, order := range orders {
-		createdEvent, ok := hashToEvent[order.TxHash]
-		if !ok {
-			continue
-		}
-
-		wg.Add(1)
-		go func(order *ent.PaymentOrder, createdEvent *types.OrderCreatedEvent) {
-			defer wg.Done()
-
-			createdEvent.Amount = createdEvent.Amount.Div(decimal.NewFromInt(10).Pow(decimal.NewFromInt(int64(order.Edges.Token.Decimals))))
-			createdEvent.ProtocolFee = createdEvent.ProtocolFee.Div(decimal.NewFromInt(10).Pow(decimal.NewFromInt(int64(order.Edges.Token.Decimals))))
-
-			err := common.CreateLockPaymentOrder(ctx, network, createdEvent, s.order.RefundOrder, s.priorityQueue.AssignLockPaymentOrder)
-			if err != nil {
-				if !strings.Contains(fmt.Sprintf("%v", err), "duplicate key value violates unique constraint") {
-					logger.WithFields(logger.Fields{
-						"Error":   fmt.Sprintf("%v", err),
-						"OrderID": createdEvent.OrderId,
-						"TxHash":  createdEvent.TxHash,
-						"Network": network.Identifier,
-					}).Errorf("Failed to create lock payment order when indexing order created events for %s", network.Identifier)
-				}
-				return
-			}
-
-			// Update payment order with txHash
-			_, err = order.Update().
-				SetGatewayID(createdEvent.OrderId).
-				SetRate(order.Rate).
-				SetStatus(paymentorder.StatusPending).
-				Save(ctx)
-			if err != nil {
-				return
-			}
-
-			// Refetch payment order
-			paymentOrder, err := storage.Client.PaymentOrder.
-				Query().
-				Where(paymentorder.IDEQ(order.ID)).
-				WithSenderProfile().
-				Only(ctx)
-			if err != nil {
-				return
-			}
-
-			// Send webhook notifcation to sender
-			err = utils.SendPaymentOrderWebhook(ctx, paymentOrder)
-			if err != nil {
-				return
-			}
-		}(order, createdEvent)
-	}
-	wg.Wait()
 
 	return nil
 }
@@ -253,44 +179,10 @@ func (s *IndexerEVM) IndexOrderSettled(ctx context.Context, rpcClient types.RPCC
 		return nil
 	}
 
-	lockOrders, err := storage.Client.LockPaymentOrder.
-		Query().
-		Where(
-			lockpaymentorder.TxHashIn(txHashes...),
-			lockpaymentorder.StatusEQ(lockpaymentorder.StatusValidated),
-		).
-		WithToken(func(tq *ent.TokenQuery) {
-			tq.WithNetwork()
-		}).
-		All(ctx)
+	err = common.ProcessSettledOrders(ctx, network, txHashes, hashToEvent)
 	if err != nil {
-		return fmt.Errorf("IndexOrderSettled.fetchLockOrders: %w", err)
+		return fmt.Errorf("IndexOrderSettled.processSettledOrders: %w", err)
 	}
-
-	var wg sync.WaitGroup
-	for _, lockOrder := range lockOrders {
-		settledEvent, ok := hashToEvent[lockOrder.TxHash]
-		if !ok {
-			continue
-		}
-
-		wg.Add(1)
-		go func(lo *ent.LockPaymentOrder, se *types.OrderSettledEvent) {
-			defer wg.Done()
-
-			// Update order status
-			err := common.UpdateOrderStatusSettled(ctx, network, se)
-			if err != nil {
-				logger.WithFields(logger.Fields{
-					"Error":   fmt.Sprintf("%v", err),
-					"OrderID": se.OrderId,
-					"TxHash":  se.TxHash,
-					"Network": network.Identifier,
-				}).Errorf("Failed to update order status settlement when indexing order settled events for %s", network.Identifier)
-			}
-		}(lockOrder, settledEvent)
-	}
-	wg.Wait()
 
 	return nil
 }
@@ -333,46 +225,10 @@ func (s *IndexerEVM) IndexOrderRefunded(ctx context.Context, rpcClient types.RPC
 		return nil
 	}
 
-	lockOrders, err := storage.Client.LockPaymentOrder.
-		Query().
-		Where(
-			lockpaymentorder.TxHashIn(txHashes...),
-			lockpaymentorder.Or(
-				lockpaymentorder.StatusEQ(lockpaymentorder.StatusPending),
-				lockpaymentorder.StatusEQ(lockpaymentorder.StatusCancelled),
-			),
-		).
-		WithToken(func(tq *ent.TokenQuery) {
-			tq.WithNetwork()
-		}).
-		All(ctx)
+	err = common.ProcessRefundedOrders(ctx, network, txHashes, hashToEvent)
 	if err != nil {
-		return fmt.Errorf("IndexOrderRefunded.fetchLockOrders: %w", err)
+		return fmt.Errorf("IndexOrderRefunded.processRefundedOrders: %w", err)
 	}
-
-	var wg sync.WaitGroup
-	for _, lockOrder := range lockOrders {
-		wg.Add(1)
-		go func(lockOrder *ent.LockPaymentOrder) {
-			defer wg.Done()
-			refundedEvent, ok := hashToEvent[lockOrder.TxHash]
-			if !ok {
-				return
-			}
-
-			refundedEvent.Fee = refundedEvent.Fee.Div(decimal.NewFromInt(10).Pow(decimal.NewFromInt(int64(lockOrder.Edges.Token.Decimals))))
-
-			err := common.UpdateOrderStatusRefunded(ctx, lockOrder.Edges.Token.Edges.Network, refundedEvent)
-			if err != nil {
-				logger.WithFields(logger.Fields{
-					"Error":   fmt.Sprintf("%v", err),
-					"OrderID": refundedEvent.OrderId,
-					"TxHash":  refundedEvent.TxHash,
-				}).Errorf("Failed to update order status refund when indexing order refunded events for %s", lockOrder.Edges.Token.Edges.Network.Identifier)
-			}
-		}(lockOrder)
-	}
-	wg.Wait()
 
 	return nil
 }
